Fix FailedWithSeed junit integration bug (https://github.com/IanEsling/fyodor/issues/23 ).
The original junit integration example invokes initCause on the lowest exception of the exception stacktrace, resulting in the FailedWithSeedException being the very bottom exception.

Due to the behaviour of Throwable#initCause(Throwable) there are times when this method will throw an IllegalStateException.

The solution is to first throw the FailedWithSeed, and set the cause of this as the original exception.

FailedWithSeed is Throwable so junit marks the test as a failure, rather than in error.